requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0
nltk>=3.8
gensim>=4.3.0
